test: Add jobmanager.service.test
